Fix search of MRI test exclusion files.
In some cases we have directories that match the search prefix and they will naturally not include the tagged test data.

+++ tool/jt.rb
@@ -1371,7 +1371,7 @@ module Commands
       test_classes.each do |test_class|
         prefix = "test/mri/excludes/#{test_class.gsub('::', '/')}"
         ["#{prefix}.rb", prefix].each do |file|
-          if File.exist?(file)
+          if File.exist?(file) && File.file?(file)
             lines = File.readlines(file)
             FileUtils::Verbose.rm_r file
